+++ modules/admin-ui/src/main/java/org/opencastproject/adminui/endpoint/ServicesEndpoint.java
@@ -211,6 +211,10 @@ public class ServicesEndpoint {
     public static final String RUNNING_NAME = "running";
     /** Status model field name. */
     public static final String STATUS_NAME = "status";
+    /** Online model field name. */
+    public static final String ONLINE_NAME = "online";
+    /** Maintenance model field name. */
+    public static final String MAINTENANCE_NAME = "maintenance";
 
     /** Wrapped {@code ServiceStatistics} instance. */
     private final ServiceStatistics serviceStatistics;
@@ -296,6 +300,22 @@ public class ServicesEndpoint {
     }
 
     /**
+     * Returns whether the service is online.
+     * @return online status
+     */
+    public boolean getIsOnline() {
+      return serviceStatistics.getServiceRegistration().isOnline();
+    }
+
+    /**
+     * Returns whether the service is in maintenance.
+     * @return maintenance status
+     */
+    public boolean getisMaintenance() {
+      return serviceStatistics.getServiceRegistration().isInMaintenanceMode();
+    }
+
+    /**
      * Returns a map of all service fields.
      * @return a map of all service fields
      */
@@ -310,6 +330,8 @@ public class ServicesEndpoint {
       serviceMap.put(QUEUED_NAME, Integer.toString(getQueuedJobs()));
       serviceMap.put(RUNNING_NAME, Integer.toString(getRunningJobs()));
       serviceMap.put(STATUS_NAME, getStatus().name());
+      serviceMap.put(ONLINE_NAME, Boolean.toString(getIsOnline()));
+      serviceMap.put(MAINTENANCE_NAME, Boolean.toString(getisMaintenance()));
       return serviceMap;
     }
 
@@ -331,7 +353,9 @@ public class ServicesEndpoint {
               f(MEAN_QUEUE_TIME_NAME, v(getMeanQueueTime())), f(MEAN_RUN_TIME_NAME, v(getMeanRunTime())),
               f(NAME_NAME, v(getName(), Jsons.BLANK)), f(QUEUED_NAME, v(getQueuedJobs())),
               f(RUNNING_NAME, v(getRunningJobs())),
-              f(STATUS_NAME, v(SERVICE_STATUS_TRANSLATION_PREFIX + getStatus().name(), Jsons.BLANK)));
+              f(STATUS_NAME, v(SERVICE_STATUS_TRANSLATION_PREFIX + getStatus().name(), Jsons.BLANK)),
+              f(ONLINE_NAME, v(getIsOnline())),
+              f(MAINTENANCE_NAME, v(getisMaintenance())));
     }
   }